[Delivers 94714266]
When a requester modifies a request that has already been approved, the approver that has already taken action gets an email. This email does not require action (i.e. we do not display an 'Approve' button), but we still display `Or Send a Comment`.  Let's change this link to say `View this request`