ensure that work orders aren't incorrectly marked as changed
Previously, direct_pay was being set in the `let` for the editing
feature specs. We took this out to make the spec setup simpler, but
revealed that submitting the edit form would change the value from `nil`
to `false`. Fixed this by setting `direct_pay = false` as a default.

Similarly, we didn't want the budget fields being left empty to change
the corresponding attributes from `nil` to empty string, so are setting
them back to `nil` in a model callback.