Add migration for ahoy tracking email
Opens

An invisible pixel is added right before the </body> tag in HTML emails.

If the recipient has images enabled in his or her email client, the pixel is loaded and the open time recorded.

Use track open: false to skip this.

Clicks

A redirect is added to links to track clicks in HTML emails.

http://chartkick.com
becomes

http://you.io/ahoy/messages/rAnDoMtOkEn/click\?url\=http%3A%2F%2Fchartkick.com\&signature\=...
A signature is added to prevent open redirects.

Use track click: false to skip tracking, or skip specific links with:

<a data-skip-click=true href=...>Can't touch this</a>